requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
azure-core==1.32.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-file-share==12.20.1
azure-storage-queue==12.12.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
Deprecated==1.2.18
fonttools==4.56.0
frozenlist==1.5.0
furl==2.1.3
idna==3.10
img2pdf==0.6.0
isodate==0.7.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
orderedmultidict==1.0.1
packaging==24.2
pandas==2.2.3
pikepdf==9.5.2
pillow==11.1.0
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
appscript==1.3.0
aspose-words==25.2.0
attrs==25.1.0
azure-core==1.32.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-file-share==12.20.1
azure-storage-queue==12.12.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.1
cssselect2==0.7.0
cycler==0.12.1
Deprecated==1.2.18
docx2pdf==0.1.8
et_xmlfile==2.0.0
fonttools==4.56.0
frozenlist==1.5.0
furl==2.1.3
idna==3.10
img2pdf==0.6.0
isodate==0.7.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.5
pikepdf==9.5.2
pillow==11.1.0
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pydyf==0.11.0
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
wrapt==1.17.2
xlrd==2.0.1
yarl==1.18.3
zopfli==0.2.3.post1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
appscript==1.3.0
aspose-words==25.2.0
attrs==25.1.0
azure-core==1.32.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-file-share==12.20.1
azure-storage-queue==12.12.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
comtypes==1.4.10
contourpy==1.3.1
cryptography==44.0.1
cssselect2==0.7.0
cycler==0.12.1
Deprecated==1.2.18
docx2pdf==0.1.8
et_xmlfile==2.0.0
fonttools==4.56.0
frozenlist==1.5.0
furl==2.1.3
idna==3.10
img2pdf==0.6.0
isodate==0.7.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.5
pikepdf==9.5.2
pillow==11.1.0
propcache==0.2.1
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pydyf==0.11.0
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdf==5.3.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
weasyprint==64.0
webencodings==0.5.1
wrapt==1.17.2
xlrd==2.0.1
yarl==1.18.3
zopfli==0.2.3.post1
